add DataFiller to upload images and data to card. Conole log works. Need to make scroll
