+++ src/pagesMinin/ModalsForTable/ModalEditCard/modalEditCardMinin.module.scss
@@ -1,0 +1,20 @@
+* p {
+  margin: 0;
+}
+.body {
+  margin-bottom: 45px;
+  position: relative;
+  z-index: 1;
+  & > :first-child {
+    margin-bottom: 24px;
+  }
+}
+
+.customClass {
+  width: 542px;
+}
+
+.footer {
+  display: flex;
+  justify-content: space-between;
+}
