+++ src/pagesMinin/featuresMinin/DeckCards.page.tsx
@@ -6,7 +6,7 @@ import Input from '@/components/ui/Input/Input'
 import { PaginationWithSelect } from '@/components/ui/Pagination/PaginationWithSelect'
 import Typography from '@/components/ui/Typography/Typography'
 import { Button } from '@/components/ui/button'
-import { ModalEditCard } from '@/pagesMinin/ModalsForTable/ModalEditCard'
+import { ModalEditCard } from '@/pagesMinin/ModalsForTable/ModalEditCard/ModalEditCard'
 import { DeckCardsTableMinin } from '@/pagesMinin/TablesMinin/DeckCardsTableMinin/DeckCardsTableMinin'
 import { PageMinin } from '@/pagesMinin/componentsMinin/PageMinin/PageMinin'
 import { useQueryParams } from '@/pagesMinin/useQueryParams'
